suppression des coms et du post

+++ backend/controllers/post-ctrl.js
@@ -1,5 +1,6 @@
 const model = require("../models");
 const modelPost = model.Post;
+const modelCom = model.Comment;
 const fs = require("fs");
 
 exports.createPost = (req, res) => {
@@ -65,15 +66,15 @@ exports.updatePost = (req, res) => {
       if (post.UserId !== req.auth.userId) {
         res.status(401).json({ message: "no autorisé" });
       } else {
-        const filename = post.attachment.split("/images/")[1];
-        fs.unlink(`images/${filename}`, () => {
-          modelPost
-            .update(newPost, { where: { id: req.params.id } })
-            .then((data) =>
-              res.status(200).json({ message: "post modifier", data })
-            )
-            .catch((err) => res.status(400).json({ err }));
-        });
+        // const filename = post.attachment.split("/images/")[1];
+        // fs.unlink(`images/${filename}`, () => {
+        modelPost
+          .update(newPost, { where: { id: req.params.id } })
+          .then((data) =>
+            res.status(200).json({ message: "post modifier", data })
+          )
+          .catch((err) => res.status(400).json({ err }));
+        // });
       }
     })
     .catch((err) => res.satus(400).json({ err }));
@@ -86,15 +87,33 @@ exports.deletePost = (req, res) => {
       if (post.UserId !== req.auth.userId) {
         res.status(401).json({ message: "pas autorisé" });
       } else {
-        const filename = post.attachment.split("/images/")[1];
-        fs.unlink(`images/${filename}`, () => {
-          modelPost
-            .destroy({ where: { id: req.params.id } })
-            .then(() => res.status(200).json({ message: "post effacé" }))
-            .catch((err) =>
-              res.status(400).json({ message: "probleme la", err })
-            );
-        });
+        modelCom
+          .findAll()
+          .then((coms) => {
+            if (coms) {
+              modelCom
+                .destroy({ where: { PostId: req.params.id } })
+                .then(() => {
+                  const filename = post.attachment.split("/images/")[1];
+                  fs.unlink(`images/${filename}`, () => {
+                    modelPost
+                      .destroy({ where: { id: req.params.id } })
+                      .then(() =>
+                        res.status(200).json({ message: "post effacé" })
+                      )
+                      .catch((err) =>
+                        res.status(400).json({ message: "probleme la", err })
+                      );
+                  });
+                })
+                .catch((err) =>
+                  res.status(400).json({ message: "probleme là", err })
+                );
+            } else {
+              console.log("bonsoir");
+            }
+          })
+          .catch((err) => res.status(400).json({ err }));
       }
     })
     .catch((err) => res.status(400).json({ message: "probleme ici", err }));
